refactor(ui): name the Mosaic UserButton layers for their MVC roles v2 (#9524)
